drop inputs, move logic to service
