+++ projects/storefrontlib/src/cms-components/myaccount/order/cancellations-returns/cancel-or-return.service.ts
@@ -1,16 +1,14 @@
 import { Injectable } from '@angular/core';
-import { formatCurrency, getCurrencySymbol } from '@angular/common';
 import {
   CancelOrReturnRequestEntryInput,
-  OrderEntry,
-  LanguageService,
-  Price,
-  RoutingService,
   GlobalMessageService,
   GlobalMessageType,
-  UserOrderService,
+  LanguageService,
+  OrderEntry,
   OrderReturnRequestService,
+  RoutingService,
   SemanticPathService,
+  UserOrderService,
 } from '@spartacus/core';
 import { Observable } from 'rxjs';
 import { take } from 'rxjs/operators';
@@ -107,27 +105,27 @@ export class OrderCancelOrReturnService {
     }
   }
 
-  /**
-   * As discussed, this calculation is moved to SPA side.
-   * The calculation and validation should be in backend facade layer.
-   */
-  getCancelledOrReturnedPrice(entry: OrderEntry): Price {
-    const qty = this.getEntryCancelledOrReturnedQty(entry);
+  // /**
+  //  * As discussed, this calculation is moved to SPA side.
+  //  * The calculation and validation should be in backend facade layer.
+  //  */
+  // getCancelledOrReturnedPrice(entry: OrderEntry): Price {
+  //   const qty = this.getEntryCancelledOrReturnedQty(entry);
 
-    const returnedItemsPriceData = Object.assign({}, entry.basePrice);
+  //   const returnedItemsPriceData = Object.assign({}, entry.basePrice);
 
-    returnedItemsPriceData.value =
-      Math.round(entry.basePrice.value * qty * 100) / 100;
+  //   returnedItemsPriceData.value =
+  //     Math.round(entry.basePrice.value * qty * 100) / 100;
 
-    returnedItemsPriceData.formattedValue = formatCurrency(
-      returnedItemsPriceData.value,
-      this.lang,
-      getCurrencySymbol(returnedItemsPriceData.currencyIso, 'narrow'),
-      returnedItemsPriceData.currencyIso
-    );
+  //   returnedItemsPriceData.formattedValue = formatCurrency(
+  //     returnedItemsPriceData.value,
+  //     this.lang,
+  //     getCurrencySymbol(returnedItemsPriceData.currencyIso, 'narrow'),
+  //     returnedItemsPriceData.currencyIso
+  //   );
 
-    return returnedItemsPriceData;
-  }
+  //   return returnedItemsPriceData;
+  // }
 
   goToOrderCancelOrReturn(cxRoute: string, orderCode: string): void {
     this.routing.go({
@@ -152,14 +150,14 @@ export class OrderCancelOrReturnService {
     return false;
   }
 
-  getEntryCancelledOrReturnedQty(entry: OrderEntry): number {
-    for (const input of this._cancelOrReturnRequestInputs) {
-      if (input.orderEntryNumber === entry.entryNumber) {
-        return input.quantity;
-      }
-    }
-    return 0;
-  }
+  // getEntryCancelledOrReturnedQty(entry: OrderEntry): number {
+  //   for (const input of this._cancelOrReturnRequestInputs) {
+  //     if (input.orderEntryNumber === entry.entryNumber) {
+  //       return input.quantity;
+  //     }
+  //   }
+  //   return 0;
+  // }
 
   cancelOrder(orderCode: string): void {
     this.userOrderService.cancelOrder(orderCode, {
