+++ projects/storefrontlib/src/cms-components/myaccount/order/cancellations-returns/cancel-order-confirmation/cancel-order-confirmation.component.html
@@ -22,7 +22,6 @@
 <ng-container *ngIf="cancelledEntries$ | async as cancelledEntries">
   <cx-cancel-or-return-items
     [entries]="cancelledEntries"
-    [cancelOrder]="true"
     [confirmRequest]="true"
   >
   </cx-cancel-or-return-items>
